Move some common solver functionality to a separate header.
